Material Dialog

+++ src/app/heroes/pages/agregar/agregar.component.ts
@@ -4,6 +4,8 @@ import { HeroesService } from '../../services/heroes.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 
 @Component({
   selector: 'app-agregar',
@@ -42,10 +44,11 @@ export class AgregarComponent implements OnInit {
   }
 
   constructor(private readonly heroesService: HeroesService,
-              private readonly activeRoute: ActivatedRoute,
-              private readonly router: Router,
-              private readonly snackBar: MatSnackBar
-    ) { }
+    private readonly activeRoute: ActivatedRoute,
+    private readonly router: Router,
+    private readonly snackBar: MatSnackBar,
+    private readonly dialog: MatDialog
+  ) { }
 
   ngOnInit(): void {
 
@@ -86,19 +89,30 @@ export class AgregarComponent implements OnInit {
 
   }
 
-  delete(){
-    if(this.heroe.id){
-      this.heroesService.deleteHeroe(this.heroe.id!)
-      .subscribe({
-        next: res => console.log(res),
-        error: () => this.router.navigate(['/heroes'])
-      })
-    }
+  delete() {
 
-    this.router.navigate(['/heroes']);
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '250px',
+      data: { ...this.heroe }
+    })
+
+    dialog.afterClosed()
+      .subscribe(res => {
+        if (res) {
+          if (this.heroe.id) {
+            this.heroesService.deleteHeroe(this.heroe.id!)
+              .subscribe({
+                next: res => console.log(res),
+                error: () => this.router.navigate(['/heroes'])
+              })
+          }
+
+          this.router.navigate(['/heroes']);
+        }
+      })
   }
 
-  showSnackBar(msj: string){
+  showSnackBar(msj: string) {
     this.snackBar.open(msj, 'Ok!', {
       duration: 2500
     })
